using int now

+++ src/Module.Server/Api/Models/Items/CrpgItem.cs
@@ -9,7 +9,7 @@ internal class CrpgItem
     public CrpgItemType Type { get; set; }
     public int Price { get; set; }
     public float Weight { get; set; }
-    public float Requirement { get; set; }
+    public int Requirement { get; set; }
     public CrpgItemArmorComponent? Armor { get; set; }
     public CrpgItemMountComponent? Mount { get; set; }
     public IList<CrpgItemWeaponComponent> Weapons { get; set; } = Array.Empty<CrpgItemWeaponComponent>();
